Moved Console.WriteLine from server to implementation.
